% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position.R
\name{add_image_additive}
\alias{add_image_additive}
\title{Add one image additively to another one at given position}
\usage{
add_image_additive(im, tgt, x, y)
}
\arguments{
\item{im}{background image}

\item{tgt}{target image that will be added to the background}

\item{x}{x position of the target}

\item{y}{y position of the target}
}
\value{
merged image array
}
\description{
Add one image additively to another one at given position
}
\examples{
# Add Rlogo image to boat image supllied with imager packager
set.seed(100)
bckg <- imager::boats  \%>\% imager::grayscale()
tgt <- imager::load.image(system.file("img", "Rlogo.png", package="png")) \%>\% imager::rm.alpha() \%>\% imager::grayscale() \%>\% imager::resize_halfXY()
plot(imager::as.cimg(add_image_additive(bckg, tgt, ncol(bckg)/2, nrow(bckg)/2)))
}
