% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{add_random_direction}
\alias{add_random_direction}
\title{Add random direction}
\usage{
add_random_direction(position)
}
\arguments{
\item{position}{tibble}
}
\value{
same tibble with extra direction column
}
\description{
Simple fuction to assign random direction to each object in position object.
Random means uniformly distributed value from 0 to 2 * pi (0 to 360 deg).
}
