% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position.R
\name{annotate_aperture}
\alias{annotate_aperture}
\title{Draw bitmap image as ggplot2 annotation and cover rest with filled colour}
\usage{
annotate_aperture(
  image,
  xlim_in,
  ylim_in,
  colour,
  alpha = 1,
  xlim_out = NULL,
  ylim_out = NULL,
  settings = NULL
)
}
\arguments{
\item{image}{rasterGrob object or filename (png)}

\item{xlim_in}{x-coordinates of the image}

\item{ylim_in}{y-coordinates of the image}

\item{colour}{fill colour for the remaining area}

\item{alpha}{Transparency for remaining area (not for the drawn image)}

\item{xlim_out}{x-coordinates of the covered area}

\item{ylim_out}{y-coordinates of the covered area}

\item{settings}{list with named parameters, default values for xlim_out and ylim_out
can be taken from xlim and ylim}
}
\value{
ggplot2 layer (as in `ggplot2::annotate`)
}
\description{
Draw bitmap image as ggplot2 annotation and cover rest with filled colour
}
\examples{
image <- system.file("extdata/aperture_white.png", package="motrack")
s <- new_settings()
plot_position(position8c, s,
  foreground_annotation = annotate_aperture(
    image, xlim_in = c(-6, 6), ylim_in = c(-6, 6), colour = "white", settings = s
))
}
