% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position.R
\name{annotate_image}
\alias{annotate_image}
\title{Draw bitmap image as ggplot2 annotation}
\usage{
annotate_image(image, xlim = NULL, ylim = NULL, settings = NULL)
}
\arguments{
\item{image}{rasterGrob object or filename (png)}

\item{xlim}{x-coordinates}

\item{ylim}{y-coordinates}

\item{settings}{list with named parameters, from which default values for xlim and ylim are taken}
}
\value{
ggplot2 layer (as in `ggplot2::annotation_custom`)
}
\description{
Draw bitmap image as ggplot2 annotation
}
