% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{bounce_off_circle}
\alias{bounce_off_circle}
\title{Check for circular arena bouncing and recalculate directions}
\usage{
bounce_off_circle(moment, time_next, settings)
}
\arguments{
\item{moment}{Position tibble with extra columns `direction` and `speed`}

\item{time_next}{Next timepoint}

\item{settings}{list with basic properties}
}
\value{
Like `moment`, but some directions are adjusted (reversed, swapped),
if there were a collision with arena border in next time point (`timestep`)
}
\description{
Check for circular arena bouncing and recalculate directions
}
