% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{bounce_pair}
\alias{bounce_pair}
\title{Calculate collsion vectors for two objects}
\usage{
bounce_pair(v1, v2, x1, x2)
}
\arguments{
\item{v1}{Velocity vector of object 1. Two-dimensional numeric vector.}

\item{v2}{Velocity vector of object 2. Two-dimensional numeric vector.}

\item{x1}{Center of object 1 in the time collision. Two-dimensional numeric vector.}

\item{x2}{Center of object 1 in the time collision. Two-dimensional numeric vector.}
}
\value{
list of the two two-dimension vectors - new speed for object 1 and 2
}
\description{
Calculate collsion vectors for two objects
}
\examples{
# direct collision, speeds swap
bounce_pair(c(0, 1), c(0, -1), c(0, 0), c(0, 1))

# 45deg contact angle, makes 90deg turns
bounce_pair(c(0, 1), c(0, -1), c(0, 0), c(1, 1))
bounce_pair(c(0, 1), c(0, -1), c(0, 1), c(1, 0))

# distance is not relevant, direct or 45 deg contact
bounce_pair(c(0, 1), c(0, -1), c(0, 0), c(0, 2))
bounce_pair(c(0, 1), c(0, -1), c(0, 0), c(2, 2))

# Works also for different velocities
bounce_pair(c(0, 1), c(0, -2), c(0, 0), c(0, 1))
bounce_pair(c(0, 1), c(0, -2), c(0, 0), c(1, 1))
}
