% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position.R
\name{convert_position_to_pixels}
\alias{convert_position_to_pixels}
\title{Converts position in degrees to pixels to have comparable length with}
\usage{
convert_position_to_pixels(position, border_length_pix, border_length_deg)
}
\arguments{
\item{position}{tibble}

\item{border_length_pix}{borders of arena in pixels - c(x,y)}

\item{border_length_deg}{length of arena in degrees}
}
\value{
tibble
}
\description{
Converts position in degrees to pixels to have comparable length with
}
\examples{
settings = new_settings()
pos <- generate_positions_random(8, settings)
convert_position_to_pixels(pos, c(750,750),sum(abs(settings$xlim)))

}
