% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position.R
\name{default_settings}
\alias{default_settings}
\title{Default settings for position/trajectory code}
\usage{
default_settings()
}
\value{
list of parameters
}
\description{
Returns a list of values used for position/trajectory generation or
presentation. The values refer to the "physical properties"
(e.g., definition of the objects or arena) or
to the presentation propoerties
(e.g., objects' colour, presence of border).
The list is passed as a parameter to other functions.
}
\details{
Currently used properties include:
\describe{
  \item{xlim}{A vector of form `c(xmin, xmax)`.
  Represents x-limit of the arena square, objects are limited to this area.}
  \item{ylim}{A vector of form `c(ymin, ymax)`.
  Represents y-limit of the arena square, objects are limited to this area.}
  \item{min_distance}{Minimum pairwise distance between centres of objects.}
  \item{r}{Radius of the object. Object is considered to be a circle.}
  \item{arena_border}{Logical. Whether arena border should be drawn.}
  \item{arena_shape}{Character. "square" or "circle"}
  \item{fill_object}{Character. Colour code of default fill colour.}
  \item{fill_target}{Character. Colour code of target fill colour.}
  \item{border_object}{Character. Colour code of default border.}
  \item{border_target}{Character. Colour code of target border.}
  \item{show_labels}{Logical. Should we show object numbers in plots?}
  \item{bounce_off_square}{Logical. Should objects bounce off square arena?}
  \item{bounce_off_others}{Logical. Should objects bounce off each other?}
  \item{simplified_bouncing}{Logical.
    Should pairwise collisions respect contact angle or
    just swap velocity vectors (simplified bouncing)? }
  \item{bounce_off_circle}{Logical. Should objects bounce off circular arena?}
  \item{circle_bounce_jitter}{Real. Amount of uniform angular jitter after bouncing, in radians.}
}
}
\examples{
default_settings()
}
