% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{extrapolate_moment}
\alias{extrapolate_moment}
\title{Extrapolates object positions based on current state and time}
\usage{
extrapolate_moment(moment, timestep, cur_time, new_time = NULL)
}
\arguments{
\item{moment}{Position tibble with extra columns `direction` and `speed`}

\item{timestep}{How far into future should the objects move (linearly)}

\item{cur_time}{Current time, for time-sensitive speed functions (e.g., accelerations)}

\item{new_time}{New time value for the returned object. If omitted, it is calculated automatically as `cur_time + timestep`}
}
\value{
Another moment - position tibble with
extra columns `direction` and `speed` corresponding to time_next
}
\description{
Intended for the use in custom step functions.
}
\examples{
moment <- position8c \%>\% add_random_direction() \%>\%
  dplyr::mutate(speed = 3)
extrapolate_moment(moment, 0.1, 0)
}
