% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position.R
\name{generate_positions_random}
\alias{generate_positions_random}
\title{Random object positions}
\usage{
generate_positions_random(n, settings, check_distance = T)
}
\arguments{
\item{n}{Number of objects}

\item{settings}{Basic properties - namely `xlim`, `ylim`,
possibly `min_distance`}

\item{check_distance}{Logical.
The positions are generated until
minimum pairwise distance of `min_dist` is met.}
}
\value{
Tibble with `object`, `x` and `y` columns.
}
\description{
The function expects a square arena defined by `xlim` and `ylim` settings.
The objects are placed randomly (uniform distribution sampling) into the arena.
If `check_distance` is `TRUE`, the process is repeated
until the minimum pairwise distance is met.
}
\examples{
pos <- generate_positions_random(8, default_settings())
}
\seealso{
\code{\link{default_settings}} for setting definitions,
\code{\link{settings}} for adjusting default settings,
\code{\link{is_distance_at_least}} for checking distances
}
