% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position.R
\name{is_distance_at_least}
\alias{is_distance_at_least}
\title{Tests inter-object distances}
\usage{
is_distance_at_least(position, min_distance)
}
\arguments{
\item{position}{tibble}

\item{min_distance}{Minimum distance to be present between each pair}
}
\value{
TRUE if all distances are larger than `min_distance`
}
\description{
The function calculated all pairwise distances and checks
if the are larger or equal.
}
\examples{
example_pos <-
  tibble::tibble(object = 1:8, x = 1:8, y = 4 - (1:8))
is_distance_at_least(example_pos, 1)
is_distance_at_least(example_pos, 2)
}
