% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{is_valid_trajectory}
\alias{is_valid_trajectory}
\title{Is the object a valid trajectory}
\usage{
is_valid_trajectory(trajectory)
}
\arguments{
\item{trajectory}{tibble}
}
\value{
TRUE if conditions above are met, FALSE otherwise
}
\description{
The function tests an object (tibble) for several requirements:
\itemize{
  \item columns `time`, `object`, `x` and `y` are present.
  \item `object` column contains unique values
  \item no value missing in `x` and `y`
  \item `x` and `y` are numeric
}
}
\details{
No restriction about data type of `object` is placed.
}
\examples{
example_pos <-
  tibble::tibble(object = 1:8, x = 1:8, y = 4 - (1:8))
is_valid_position(example_pos)
}
