% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{load_trajectory}
\alias{load_trajectory}
\title{Load trajectory from file}
\usage{
load_trajectory(filename, delim = ",", ...)
}
\arguments{
\item{filename}{Name of input file (e.g., .csv)}

\item{delim}{Single character used to separate fields within a record. Default `","`}

\item{...}{Parameters passed to `readr::read_delim`}
}
\description{
Load trajectory from file
}
