% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{make_random_trajectory}
\alias{make_random_trajectory}
\title{Create random trajectory}
\usage{
make_random_trajectory(start, timescale, settings, step_function, ...)
}
\arguments{
\item{start}{Position tibble with starting configuration}

\item{timescale}{Numeric vector with time points for the trajectory}

\item{settings}{list with basic properties}

\item{step_function}{function which is called in every time step
to calculate next position in sequence}

\item{...}{Extra parameters passed to `step_function`}
}
\value{
Tibble trajectory object
}
\description{
Create random trajectory
}
\examples{
sett_move <-
  new_settings(
    speed = 5, xlim = c(-9, 9), ylim = c(-9, 9),
    bounce_off_square = FALSE,
    bounce_off_circle = TRUE, circle_bounce_jitter = pi / 6
  )
moment <- position8c \%>\% add_random_direction()
tt <- make_random_trajectory(
  moment, seq(0, 8, by = 0.1), sett_move, step_direct
)
}
