% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position.R
\name{new_settings}
\alias{new_settings}
\title{Creates a custom copy of settings}
\usage{
new_settings(...)
}
\arguments{
\item{...}{named list of properties to override default settings.}
}
\value{
Updated list of settings
}
\description{
Creates a custom copy of settings
}
\examples{
new_settings(xlim = c(0, 10), ylim = c(0, 10))
}
\seealso{
\code{\link{default_settings}}
}
