% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position.R
\name{plot_position_image}
\alias{plot_position_image}
\title{Plot object positions by adding targets to the background}
\usage{
plot_position_image(position, settings = default_settings(),
  targets = NULL)
}
\arguments{
\item{position}{tibble}

\item{settings}{list with basic properties}

\item{targets}{Which objects should be treated as targets}
}
\value{
ggplot2 figure
}
\description{
The function plots position tibble based on x, y values and potentially
extra values from settings. Position is expected to be position tibble,
but it may contain extra columns for graphics:
target (Logical vector indicating whether object is target or distractor),
fill (Character vector with colour names of object interiors.),
border (Character vector with colour names of object borders.).
}
\details{
This function requires two parameters to be present in the settings:
target_img (image that will be added additively to the background)
background (image that will be added to the background)
Additionally, it is also good to overwrite the default target and distractor borders and fill
}
\examples{
# sample positions with no other requirements
set.seed(100)
bckg <- imager::boats
tgt <- imager::load.image(system.file("img", "Rlogo.png", package="png"))
settings_additive <- new_settings(background = bckg, target_img = tgt, fill_target = NA, fill_object = NA, border_object = NA, border_target = "green")
pos <- generate_positions_random(8, settings_additive)
plot_position_image(pos, settings_additive)
# first four objects are targets
plot_position(pos, default_settings(), 1:4)
pos$fill <- rainbow(8)
plot_position(pos, default_settings())
}
