% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{plot_trajectory}
\alias{plot_trajectory}
\title{Plot object trajectories}
\usage{
plot_trajectory(trajectory, settings = default_settings(), targets = NULL, ...)
}
\arguments{
\item{trajectory}{tibble}

\item{settings}{list with basic properties}

\item{targets}{Which objects should be treated as targets}

\item{...}{Parameters passed to `plot_position`}
}
\value{
ggplot2 figure
}
\description{
The function plots trajectory tibble based on x, y,
time values and potentially
extra values from settings.
trajectory is expected to be trajectory tibble,
but it may contain extra columns for graphics (same as plot_position):
target (Logical vector indicating whether object is target or distractor),
fill (Character vector with colour names of object interiors.),
border (Character vector with colour names of object borders.).
}
\examples{
plot_trajectory(trajectory8c, default_settings())

# first four objects are targets
plot_trajectory(trajectory8c, default_settings(), 1:4)
}
