% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{render_trajectory_video}
\alias{render_trajectory_video}
\title{Render video from trajectory}
\usage{
render_trajectory_video(filename, trajectory,
  settings = default_settings(), targets = NULL, outdir = getwd())
}
\arguments{
\item{filename}{the file name of output video (e.g. "trajectory.mp4")}

\item{trajectory}{tibble with trajectory data}

\item{settings}{list with basic properties}

\item{targets}{Which objects should be treated as targets}

\item{outdir}{output directory where the video is saved}
}
\value{
No return value
}
\description{
The function repeatedly calls plot_position function to generate preview
of the trajectory motion.
Extra time is added before and after the motion sequence (2 seconds each).
}
\examples{
# set ffmpeg in necessary

# ani.options(ffmpeg = "/PATH/TO/ffmpeg/ffmpeg")
render_trajectory_video("trajectory.mp4", trajectory8c,
  new_settings(show_labels = T), targets = 1:4)
}
