% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{save_trajectory}
\alias{save_trajectory}
\title{Save trajectory to file}
\usage{
save_trajectory(trajectory, filename, delim = ",")
}
\arguments{
\item{trajectory}{Trajectory object to be exported to file}

\item{filename}{Name of output file (e.g., .csv)}

\item{delim}{Single character used to separate fields within a record. Default `","`}
}
\description{
Save trajectory to file
}
\examples{
filename <- "test-csv"
save_trajectory(trajectory8c, filename)
unlink(filename)
}
