% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position.R
\name{settings}
\alias{settings}
\title{Creates a custom copy of settings}
\usage{
settings(...)
}
\arguments{
\item{...}{named list of properties to override default settings.}
}
\value{
Updated list of settings
}
\description{
Creates a custom copy of settings
}
\examples{
settings(xlim = c(0, 10), ylim = c(0, 10))
}
\seealso{
\code{\link{default_settings}}
}
