% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{step_direct}
\alias{step_direct}
\title{Simple trajectory step function}
\usage{
step_direct(moment, time_next, settings)
}
\arguments{
\item{moment}{Position tibble with extra columns `direction` and `speed`}

\item{time_next}{Next time step to predict}

\item{settings}{list with basic properties}
}
\value{
Another moment - position tibble with
extra columns `direction` and `speed` corresponding to time_next
}
\description{
Simple trajectory step function
}
