% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{step_vonmises}
\alias{step_vonmises}
\title{Von Mises trajectory step function}
\usage{
step_vonmises(moment, time_next, settings, kappa)
}
\arguments{
\item{moment}{Position tibble with extra columns `direction` and `speed`}

\item{time_next}{Next time step to predict}

\item{settings}{list with basic properties}

\item{kappa}{concentration parameter}
}
\value{
Another moment - position tibble with
extra columns `direction` and `speed` corresponding to time_next
}
\description{
Von Mises trajectory step function
}
