% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{step_waitandmove}
\alias{step_waitandmove}
\title{Wait and move step function}
\usage{
step_waitandmove(
  moment,
  time_next,
  settings,
  move_time = c(0.5, 1.5),
  wait_time = c(0.2, 0.5),
  wait_prob = 0.1,
  syncstart = F
)
}
\arguments{
\item{moment}{Position tibble with extra columns `direction` and `speed`}

\item{time_next}{Next time step to predict}

\item{settings}{List with basic properties}

\item{move_time}{Range of move phase duration}

\item{wait_time}{Range of wait phase duration}

\item{wait_prob}{Probability of waiting}

\item{syncstart}{Logical, if time-to-change should be synchronized
when it is generated for the first time}
}
\value{
Another moment - position tibble with
extra columns `direction`, `speed`, `speed_scale`, `ttc` and `mov_state`
}
\description{
Step function for wait-and-move movement pattern
}
\details{
Using this step function, objects changes between two states - moving and waiting.
After each state, new state is sampled with given proabability.

When probability of waiting is set to zero, objects move in a same way as in  \link{step_zigzag} step function
}
\seealso{
Other step_functions: 
\code{\link{step_direct}()},
\code{\link{step_vonmises}()},
\code{\link{step_zigzag}()}
}
\concept{step_functions}
