% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{step_zigzag}
\alias{step_zigzag}
\title{Zigzag trajectory step function}
\usage{
step_zigzag(moment, time_next, settings, ttt = c(0.5, 1.5),
  syncstart = F)
}
\arguments{
\item{moment}{Position tibble with extra columns `direction` and `speed`}

\item{time_next}{Next time step to predict}

\item{settings}{list with basic properties}

\item{ttt}{Time-to-turn - time when the object should pick new direction}

\item{syncstart}{Logical, if time-to-turn should be synchronized
when it is generated for the first time}
}
\value{
Another moment - position tibble with
extra columns `direction`, `speed` and `ttt` corresponding to time_next
}
\description{
Zigzag trajectory step function
}
