% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trajectory8c}
\alias{trajectory8c}
\title{Example trajectory}
\format{
A tibble with 648 rows and 7 variables:
\describe{
  \item{time}{Time for each position, from 0 to 8 with 0.1 steps}
  \item{object}{Identifier of each object, 1 to 8}
  \item{x}{x-coordinate, in deg}
  \item{y}{y-coordinate, in deg}
  \item{direction}{Current direction, in rad}
  \item{speed}{Object speed, in deg/s}
  \item{time1}{?}
}
}
\source{
Generated with script
}
\usage{
trajectory8c
}
\description{
Eight objects in circular arena (radius = 10, speed = 5, 8 seconds).
}
\keyword{datasets}
