% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{validate_object_direction_change}
\alias{validate_object_direction_change}
\title{Check direction changes for single object trajectory}
\usage{
validate_object_direction_change(trajectory1, change_angle = 30,
  min_interval = 0.2)
}
\arguments{
\item{trajectory1}{Trajectory featuring only one object}

\item{change_angle}{Angle value in degrees.
Direction changes beyond this value are considered large.}

\item{min_interval}{Time in seconds, which should pass between
too large direction changes.}
}
\value{
Logical. `TRUE` if each pair of consecutive bounces
is separated by at least `min_interval` seconds.
}
\description{
Check direction changes for single object trajectory
}
