% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{validate_trajectory}
\alias{validate_trajectory}
\title{Sanity check for trajectory}
\usage{
validate_trajectory(trajectory, change_angle = 30, min_interval = 0.2,
  min_distance = 1, verbose = F)
}
\arguments{
\item{trajectory}{Tibble object with trajectory}

\item{change_angle}{Angle value in degrees.
Direction changes beyond this value are considered large.}

\item{min_interval}{Time in seconds, which should pass between
too large direction changes.}

\item{min_distance}{Minimum inter-object spacing, which should be
complied in every frame.}

\item{verbose}{If `TRUE`, results for each subtest are printed on console.
Default `FALSE`.}
}
\value{
Logical. `TRUE` if all conditions are passed.
}
\description{
Sometimes, we want have extra requirements for trajectories. The idea is to
have automated check about possible errors or artifacts which could appear
in trajectory data.
}
\details{
We test for following conditions:
\itemize{
  \item *Sudden motion changes* - Sometimes objects bounce too frequently,
  which may attract attention. For each object, we test what is the minimal
  time interval between two consecutive large direction changes.
  \item *Minimum interobject distance* - The minimum distance is checked
  when generating positions and also in each step of generating trajectories.
  Here, we make final check if the minimum distance is complied.
}
}
\examples{
validate_trajectory(trajectory8c)
}
