% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{make_random_trajectory}
\alias{make_random_trajectory}
\title{Create random trajectory}
\usage{
make_random_trajectory(start, timescale, settings, step_function)
}
\arguments{
\item{start}{Position tibble with starting configuration}

\item{timescale}{Numeric vector with time points for the trajectory}

\item{settings}{list with basic properties}

\item{step_function}{function which is called in every time step
to calculate next position in sequence}
}
\description{
Create random trajectory
}
