% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{render_trajectory_video_additive}
\alias{render_trajectory_video_additive}
\title{Render video from trajectory by additive adding target images to the background}
\usage{
render_trajectory_video_additive(filename, trajectory,
  settings = default_settings(), targets = NULL, outdir = getwd())
}
\arguments{
\item{filename}{the file name of output video (e.g. "trajectory.mp4")}

\item{trajectory}{tibble with trajectory data}

\item{settings}{list with basic properties}

\item{targets}{Which objects should be treated as targets}

\item{outdir}{output directory where the video is saved}
}
\value{
No return value
}
\description{
The function repeatedly calls plot_position_additive function to generate preview
of the trajectory motion.
This function differs from render_trajectory_video as it uses creates plot positions by adding targets to background additively (which can't be done using simple ggplot logic)
Extra time is added before and after the motion sequence (2 seconds each).
}
\examples{
# set ffmpeg in necessary

# ani.options(ffmpeg = "/PATH/TO/ffmpeg/ffmpeg")
render_trajectory_video_additive("trajectory.mp4", trajectory8c,
  new_settings(show_labels = T),
  targets = 1:4
)
}
